Create CityBikeLaneData
This bike lane table includes information such as street names, year lanes were first installed, year they were updated, lane width, and safety rating. There are multiple entries for some bike lanes due to two separate technicians rating the lanes. The purpose of this analysis was to do the following:
Provide a list of all the bike lanes that have an average safety rating of 4.0 or higher. I wanted to show the rating and label it as “Safe Lane.” 
Categorize all bike lanes so that recommendations could be made as to which lanes should be removed, improved, or left as-is.